+++ frontend/src/app/box-window/box-window.component.scss
@@ -1,0 +1,3 @@
+ion-title{
+  color: yellow;
+}
